Add initial support for cri sandboxing
This is really nothing special and only partially works.
The way containerd's cri module works is it creates containers with some
special annotations, most importantly a sandbox ID.
The shim uses that sandbox ID do determine how to group containers to a
single shim.

It also uses a "pause" container as its sandbox, but the shim is
completely unaware of this, except that it is the first container
created within a sandbox grouping.
This tries to take advantage of that to be able to deal with these pause
containers.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>