Clean-up wasi setup code now that fork is a thing.
Since we're using fork now there is no need to spin up a new thread for
most of the module setup. We only need a new thread to wait on the wasi
function call to return to report the status.

We also store the pidfd and use that for sending signals to.