WIP: Store compiled modules
Modules are compiled and stored in a content addressable store.
Two runs of the same module are significantly faster since we only need
to digest the wasm module being loaded and do not need to compile it
again.

This is still a work in progress.
The store is not really concurrency safe yet (it mostly works, but multiple
threads/processes could hit issues).

Signed-off-by: Brian Goff <cpuguy83@gmail.com>