implements can_handle in wasmtime executor.
This commit implements can_handle in wasmtime executor. It checks
if the entrypoint binary has wasm as the path extension. If not,
the wasmtime executor will not be used.

It adds a default executor to the wasmtime shim as a backup
so it enables running wasm and containers side-by-side.

Add wat extension support in can_handle

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>