test: add a python container to wasmtime shim for testing
this commit adds a python container in the test directory.
It also adds a few commands in the makefile to build the python
flask app into a container, load into the kind cluster for testing.
Eventually the wasmtime tests will have both python container and
a wasm contaienr running in a pod.

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>