ci(*): automate the version bump process
this commit automates the version bump process to
ask the bot to create a PR to bump the respective
crate's version to the given one on behalf of humans.

the PR created by the bot contains magic message that
will be parsed by another workflow `publish` which triggers
on merge to the main branch.

The `publish` workflow parses the commit message to get
crate, version and dry-run information to be able to
publish the crate, release binary and push tags.

the motivation for this commit is to further simplify the
release process

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>