ci(*): remove signing from the build CI
this commit removes signing from the build CI as the release
pipeline will now handle signing of the binaries downloaded
using actions/download-artifact.

the biggest change is that the parse script is re-write to
be able to output all the necessary variables such as
`runtime`, `dry_run` and more.

the action-sign workflow is also modified to take an extra
`path` input to replace the hard-coded `dist/bin` path

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>