feat(shim): change the current dir back to bundle path
This commit changes the current dir for the shim process back
to the bundle path (e.g. `/run/containerd/io.containerd.runtime.v2.task/<ns>/<id>`)
after `shim::wait()` finishes.

The motivation for this change is that the `create` request handled by
youki's libcontainer will change the current dir to container path (e.g. `/run/containerd/wasmtime/<ns>/<id>`),
this path does not contain the `address` file for getting the shim socket address that needs
to be deleted by the ttrpc server closes.

So in order to properly delete the socket path, the current directory used by
the shim crate has to be the bundle path.

Signed-off-by: jiaxiao zhou <jiazho@microsoft.com>