chore: switch to wasm32-wasip1 target
Rust 1.81.0 now issues a warning for the usage of wasm32-wasi target and suggests
users to switch to wasm32-wasip1 target. Since we have bumped the minimal rustc version
to 1.81.0 in rust-toolchain.toml, we should switch to this new target.

This commit does a whole replace of 'wasm32-wasi -> wasm32-wasip1'.

Signed-off-by: Jiaxiao (mossaka) Zhou <duibao55328@gmail.com>