.github/workflows/action-test-smoke: Add a step to verify the jeager
Add a GitHub Action step to verify that Jeager actually receives the expected emitted traces, rather than just hoping they made it.

It uses Jeager API to fetch the traces, and note that the limit=0 so that we can get all the operations.

Signed-off-by: Jiaxiao (mossaka) Zhou <duibao55328@gmail.com>

scripts/verify-jaeger-traces: Add a script to verify jeager traces

Signed-off-by: Jiaxiao (mossaka) Zhou <duibao55328@gmail.com>

rename scripts/setup-otel.sh -> scripts/setup-jeager-and-otel.sh

Signed-off-by: Jiaxiao (mossaka) Zhou <duibao55328@gmail.com>